Clarify coding conventions
I have been interpreting "an upper bound of 120" as exactly 120. So all of my comments are wrapped at 120. I notice that there's an implicit rule to actually use 100 characters. This makes things look ugly when different authors have wrapped things differently in the same file. As well, files should end with a newline.

@randomir, @arcondello, @orenshk